Secure-Document-Platform/verification-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-jose
requests